pyproject.toml
[project]
name = "jic"
description = "[J]AX [I]ntent [C]lassification"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [{name = "JIC authors"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
]
keywords = ["speech", "jax"]

# pip dependencies of the project
dependencies = [
  "einops",
  "flax>=0.7.0",
  "last-asr==0.0.3",
  "jax",
  "optax",
]

# This is set automatically by flit using `jic.__version__`
dynamic = ["version"]

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install .[dev]`
dev = [
  "numpy",
  "pytest",
  "pytest-xdist",
  "pylint>=2.6.0",
  "yapf",
]

[build-system]
requires = ["flit_core >=3.5,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "jic"
